Changed up the file structure.
Removed the frankolang modules. Frankolang is no longer needed by Frankocoin
Core and is instead run separetly. Frankolang has been moved to a different
repoistory because of this

Added the merkle_tree crate.

Blocks module is incomplete, and doesn't have a test yet. This is because there
is nothing to test as it is incomplete.

The code doesn't do much right now, because efforts have been focused on getting
the frankolang interpreter to work.